src: remove redundant casts
All of them are implicitly convertable without any wanted side effects.

+++ extensions/libip6t_LOG.c
@@ -91,7 +91,7 @@ parse_level(const char *level)
 				   "log-level `%s' unknown", level);
 	}
 
-	return (u_int8_t)lev;
+	return lev;
 }
 
 #define IP6T_LOG_OPT_LEVEL 0x01
