+++ xtables.c
@@ -262,7 +262,7 @@ u_int16_t parse_port(const char *port, const char *proto)
 
 	if ((string_to_number(port, 0, 65535, &portnum)) != -1 ||
 	    (portnum = service_to_port(port, proto)) != (unsigned)-1)
-		return (u_int16_t)portnum;
+		return portnum;
 
 	exit_error(PARAMETER_PROBLEM,
 		   "invalid port/service `%s' specified", port);
